# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Backend Core API",
        "description": "Learning Management System Backend API Documentation. This API provides comprehensive endpoints for managing courses, users, enrollments, and learning progress.",
        "contact": {
            "name": "LMS API Support",
            "email": "api-support@lms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "LMS API Server"
        }
    ],
    "paths": {
...
        "/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get all courses",
                "description": "Retrieve a paginated list of courses for the current tenant",
                "operationId": "75403048c36d79fe464ce4716c204c61",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filter by course level",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "beginner",
                                "intermediate",
                                "advanced"
                            ]
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_free",
                        "in": "query",
                        "description": "Filter by free courses",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title and description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Courses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Courses retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course details",
                "description": "Retrieve detailed information about a specific course",
                "operationId": "703f82b2a4cf2b6bec94909bbbc1ea7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/courses/{id}/enroll": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Enroll in a course",
                "description": "Enroll the authenticated user in a course",
                "operationId": "7abbd560f122a0d11c551238ecab6b87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrolled in course successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Enrolled in course successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already enrolled or course not available"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/courses/my-courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get user's enrolled courses",
                "description": "Retrieve courses the authenticated user is enrolled in",
                "operationId": "c4d6b71d0514dc0c1a9ad8c859734546",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by enrollment status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "completed",
                                "dropped",
                                "expired",
                                "suspended"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enrolled courses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Enrolled courses retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/courses/featured": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get featured courses",
                "description": "Retrieve featured courses for the current tenant",
                "operationId": "d7109f258b8483bb0ab98466f689ec26",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of courses to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Featured courses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Featured courses retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{courseId}/enroll": {
            "post": {
                "tags": [
                    "Enrollments"
                ],
                "summary": "Enroll in a course",
                "description": "Enroll the authenticated user in a specific course",
                "operationId": "2a4bed1da07f472ffa16129ca837388c",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully enrolled in course",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully enrolled in course"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "enrolled_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "progress_percentage": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 0
                                                },
                                                "course": {
                                                    "$ref": "#/components/schemas/Course"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already enrolled or course not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You are already enrolled in this course"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "title": "Course",
                "description": "Course model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "slug": {
                        "type": "string",
                        "example": "introduction-to-laravel"
                    },
                    "short_description": {
                        "type": "string",
                        "example": "Learn Laravel framework basics"
                    },
                    "description": {
                        "type": "string",
                        "example": "Comprehensive course covering Laravel fundamentals..."
                    },
                    "thumbnail_url": {
                        "type": "string",
                        "example": "https://example.com/thumbnails/course1.jpg"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "beginner",
                            "intermediate",
                            "advanced"
                        ],
                        "example": "beginner"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "published"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "estimated_duration_hours": {
                        "type": "integer",
                        "example": 20
                    },
                    "language": {
                        "type": "string",
                        "example": "en"
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_free": {
                        "type": "boolean",
                        "example": false
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "instructor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "enrollment_count": {
                        "type": "integer",
                        "example": 150
                    },
                    "total_chapters": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}